shop_online/requirements.txt
sendgrid-django==4.2.0
